+++ Android.bp
@@ -13457,6 +13457,7 @@ filegroup {
     name: "perfetto_src_trace_processor_dataframe_dataframe",
     srcs: [
         "src/trace_processor/dataframe/dataframe.cc",
+        "src/trace_processor/dataframe/typed_cursor.cc",
     ],
 }
