+++ src/trace_processor/dataframe/BUILD.gn
@@ -18,9 +18,13 @@ source_set("dataframe") {
   sources = [
     "adhoc_dataframe_builder.h",
     "cursor.h",
+    "cursor_impl.h",
     "dataframe.cc",
     "dataframe.h",
     "runtime_dataframe_builder.h",
+    "typed_cursor.cc",
+    "typed_cursor.h",
+    "typed_cursor_impl.h",
   ]
   deps = [
     "../../../gn:default_deps",
